chore: Update Course model to include 'read' field
The Course model in models.py was updated to include a 'read' field. This field is a boolean value that indicates whether the course has been read by the user. The default value is set to False.Also Fixed bug in course.jsx's approved and reject.Extracted Searchbar in appbar.lastly implemented the notification badge in appbar